# config/swagger.json
{
    "swagger": "2.0",
    "info": {
       "description": "",
       "version": "1.0.0",
       "title": "Swagger Petstore",
       "termsOfService": "http://swagger.io/terms/",
       "contact": {
          "email": "apiteam@swagger.io"
       },
       "license": {
          "name": "Apache 2.0",
          "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
       }
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
...
    "schemes": [
       "https",
       "http"
    ],
    "paths": {
...
       "/store/order": {
          "post": {
             "tags": [
                "store"
             ],
             "summary": "Place an order for a pet",
             "description": "",
             "operationId": "placeOrder",
             "produces": [
                "application/xml",
                "application/json"
             ],
             "parameters": [
                {
                   "in": "body",
                   "name": "body",
                   "description": "order placed for purchasing the pet",
                   "required": true,
                   "schema": {
                      "$ref": "#/definitions/Order"
                   }
                }
             ],
             "responses": {
                "200": {
                   "description": "successful operation",
                   "schema": {
                      "$ref": "#/definitions/Order"
                   }
                },
                "400": {
                   "description": "Invalid Order"
                }
             }
          }
       },
...
    },
    "securityDefinitions": {
       "petstore_auth": {
          "type": "oauth2",
          "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
          "flow": "implicit",
          "scopes": {
             "write:pets": "modify pets in your account",
             "read:pets": "read your pets"
          }
       },
       "api_key": {
          "type": "apiKey",
          "name": "api_key",
          "in": "header"
       }
    },
    "definitions": {
       "Order": {
          "type": "object",
          "properties": {
             "id": {
                "type": "integer",
                "format": "int64"
             },
             "petId": {
                "type": "integer",
                "format": "int64"
             },
             "quantity": {
                "type": "integer",
                "format": "int32"
             },
             "shipDate": {
                "type": "string",
                "format": "date-time"
             },
             "status": {
                "type": "string",
                "description": "Order Status",
                "enum": [
                   "placed",
                   "approved",
                   "delivered"
                ]
             },
             "complete": {
                "type": "boolean",
                "default": false
             }
          },
          "xml": {
             "name": "Order"
          }
       },
...
    },
...
 }